update plunkr example with one that does flash empty <inputs> that say 'required' when editing
in your original plunkr, when you click to edit a project it transitions to the 'edit' page immediately (before the request to fetch the resource from the API comes back) so it would have blank inputs for a period of time and have red 'required' next to each (until the request comes back). this changes the plunker to move it up into the 'resolve' hook in $routeProvider so it waits for the request to respond before rendering the edit action so you don't have the periodic flash of blank inputs.  I think it makes the code look more clean too.